Round setup and new look
- Added packet uploading: 
Apparently to allow uploading multiple files at once you have to use the
JFileChooser, which looks teeny on high res computers. I tried to
mitigate it but it still looks kind of strange -- I'll look in to ways
to get around this but at any rate it is working. Packets are stored in
the stats object so we can access them later. At some point we will want
to go back and make sure it's checking that the file uploaded is a valid
packet. 

- Added team/player input: 
I had to add a couple of methods to your Team and Player objects to make
this work. I made new copies of them and put them in my package so we'll
want to keep that in mind when we are combining our work. 
All teams and players are saved in the stats object. If there are any
pre-existing teams, they appear as options in the ChoiceBoxes and the
textFields will auto-populate with team member names. Stats is saving
the current round and which teams are playing, so it should be pretty
easy to access this data from your part. 

- Changed to dark look